Generate commands to Vec for graphical preview
A graphical preview of the commands I am sending will make it easier for
me to catch bugs in my controller code. To display the commands, I need a
way to intercept them before they are sent, and placing them in a buffer
is an easy way to do that.