Fix bug where motors 2 & 3 went backwards
To give some background on the math, we are projecting the joystick vector
onto the motors. The motor's vectors are determined ahead of time based on
their orientation. `dot()`ting the input vector with the orientation
vector gives us a magnitude for scaling the orientation vector. More
formally;

```
s_throttle = v_input.dot(v_orientation)
s_throttle = v_input[0] * v_orientation[0] + v_input[1] * v_input[1]
```

The math is simple. Picking the correct orientation vectors, however, is
not.

Debugging the logic is even harder, because I am displaying the throttle
values on a graphics window that uses a different coordinate system then
the ROV. I didn't know if the input was being transformed incorrectly or
if the throttle value was being rendered incorrectly.

Digging into the math made it clearer.

Figuring out the vector of the line was done with this formula:

```
v_line_end = v_graphics_orientation * s_throttle
```

Again we need a magical orientaion vector to make it work. It just scales
the vector to be the size of throttle.

In this case, motor_2 had an incorrect vector. It was going up and to the
right when it should have been doing down and to the left.
```
Given:
  v_input                = < 0.0, -1.0> // forward
  v_orientation          = <-0.5,  0.5>
  v_graphics_orientation = < 0.5, -0.5>
Find v_line_end:
  v_line_end = v_graphics_orientation * s_throttle

  s_throttle = v_input[0] * v_orientation[0] + v_input[1] * v_input[1]
  s_throttle = (0)(-0.5) + (-1)(0.5)
  s_throttle = 0.5

  v_line_end = <0.5, -0.5> * (0.5)
  v_line_end = <0.25, -0.25>
```

Urgh... Sorry. This is getting too long. Long story short, I got
`v_orientation` wrong.