Remove SDL font from engine; fix font rendering
The next step is to remove all mentions of SDL_ttf from the code base.

There were a couple of errors with the text rendering. First, the outlines of
the text was blue. That was fixed by reversing the order of colors. Probably
some endianess thing... which means different platforms might need handling.
Crap. I'll deal with it then.

The second error was transparency not working. This was as simple as enabling
alpha blending for the render target and the cache texture.